Rename files for more consistency
