+++ src/pages/survey-page/form.tsx
@@ -2,12 +2,11 @@ import React, {useEffect} from 'react';
 import {types} from 'types';
 import {connect} from 'react-redux';
 import TimePill from 'components/time-pill/time-pill';
-import {pathUtils, backend} from 'utilities';
+import {pathUtils, backend, reduxUtils} from 'utilities';
 import Button from 'components/button/button';
-import SurveyField from 'components/survey-field';
+import SurveyField from 'components/survey-field/survey-field';
 import {filter} from 'lodash';
 import {useHistory} from 'react-router-dom';
-import reduxUtils from '../../utilities/redux-utils/index';
 
 function SurveyFormPage(props: {
     formConfig: types.SurveyConfig | undefined;
